requirements.txt
altgraph==0.10.2
aniso8601==9.0.1
asn1crypto==0.24.0
backports.functools-lru-cache==1.5
bdist-mpkg==0.5.0
beautifulsoup4==4.8.1
certifi==2019.11.28
cffi==1.14.3
chardet==3.0.4
click==7.1.2
configparser==4.0.2
contextlib2==0.6.0.post1
cryptography==2.6.1
decorator==4.4.2
enum34==1.1.10
Flask==1.1.2
Flask-MySQL==1.5.2
Flask-RESTful==0.3.8
Flask-SQLAlchemy==2.5.1
Flask-Migrate==2.0.2
Flask-Login==0.4.0
Flask-WTF==0.14.2
Flask-Cors==3.0.7
future==0.17.1
idna==2.8
importlib-metadata==2.1.2
ipaddress==1.0.22
ipython-genutils==0.2.0
itsdangerous==1.1.0
Jinja2==2.11.3
jupyter-core==4.6.3
macholib==1.5.1
MarkupSafe==1.1.1
#matplotlib==1.3.1
modulegraph==0.10.4
mysql-connector==2.2.9
mysql-connector-python==8.0.27
#numpy==1.8.0rc1
pathlib2==2.3.6
protobuf==3.17.3
py2app==0.7.3
pycparser==2.19
PyJWT==1.7.1
PyMySQL==0.9.3
pyOpenSSL==19.0.0
pyparsing==2.0.1
python-dateutil==1.5
pytz==2013.7
requests==2.22.0
selenium==3.141.0
six==1.12.0
soupsieve==1.9.4
SQLAlchemy==1.4.27
traitlets==4.3.3
urllib3==1.25.7
Werkzeug==0.16.1
xattr==0.6.4
zipp==1.2.0

